+++ src/main/java/com/armilp/ezvcsurvival/events/SoundEventHandler.java
@@ -13,21 +13,14 @@ public class SoundEventHandler {
 
     @SubscribeEvent
     public static void onPlaySound(PlaySoundEvent event) {
-        // Check that the sound is an instance of SimpleSoundInstance (which has a position)
         if (!(event.getSound() instanceof SimpleSound)) {
             return;
         }
 
         SimpleSound sound = (SimpleSound) event.getSound();
 
-        // Extract the position of the sound
-        double x = sound.getX();
-        double y = sound.getY();
-        double z = sound.getZ();
-        Vector3d position = new Vector3d(x, y, z);
-
-        // Get the identifier of the sound
-        ResourceLocation id = sound.getLocation();
-        SoundEventTracker.registerSound(id, position);
+        Vector3d position = new Vector3d(sound.getX(), sound.getY(), sound.getZ());
+        ResourceLocation soundId = sound.getLocation();
+        SoundEventTracker.registerSound(soundId, position);
     }
 }