+++ src/main/java/com/armilp/ezvcsurvival/goals/injector/ReactToSoundGoalInjector.java
@@ -1,6 +1,7 @@
 package com.armilp.ezvcsurvival.goals.injector;
 
-import com.armilp.ezvcsurvival.config.VoiceConfig;
+import com.armilp.ezvcsurvival.config.SoundConfig;
+import com.armilp.ezvcsurvival.data.SoundGroupData;
 import com.armilp.ezvcsurvival.goals.ReactToSoundGoal;
 import net.minecraft.entity.MobEntity;
 import net.minecraft.util.ResourceLocation;
@@ -9,7 +10,6 @@ import net.minecraftforge.event.entity.EntityJoinWorldEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 
@@ -19,17 +19,16 @@ public class ReactToSoundGoalInjector {
     @SubscribeEvent
     public static void onEntityJoinWorld(EntityJoinWorldEvent event) {
         if (!(event.getEntity() instanceof MobEntity)) {
-            return; // Verifica si la entidad es un AnimalEntity
+            return;
         }
 
         MobEntity mob = (MobEntity) event.getEntity();
 
-        Map<String, Map<String, Object>> configs = VoiceConfig.getSoundReactionConfigs();
-        ResourceLocation mobId = Registry.ENTITY_TYPE.getKey(mob.getType());
+        ResourceLocation mobIdRL = Registry.ENTITY_TYPE.getKey(mob.getType());
+        String mobId = mobIdRL.toString();
 
-        if (configs.containsKey(mobId.toString())) {
-            Map<String, Object> config = configs.get(mobId.toString());
-
+        Map<String, Object> config = SoundConfig.getMobSoundReaction(mobId);
+        if (config != null) {
             double speed = config.get("speed") instanceof Number
                     ? ((Number) config.get("speed")).doubleValue()
                     : 1.0;
@@ -37,14 +36,12 @@ public class ReactToSoundGoalInjector {
                     ? ((Number) config.get("range")).doubleValue()
                     : 16.0;
             int range = (int) rangeDouble;
-
-            @SuppressWarnings("unchecked")
-            List<String> soundTypes = config.get("sound_types") instanceof List<?>
-                    ? (List<String>) config.get("sound_types")
-                    : Collections.emptyList();
-
-            if (!soundTypes.isEmpty()) {
-                mob.goalSelector.addGoal(3, new ReactToSoundGoal(mob, speed, range, soundTypes));
+            List<?> groups = (List<?>) config.get("groups");
+            if (groups != null && !groups.isEmpty()) {
+                List<SoundGroupData> soundGroups = SoundConfig.getSoundGroupsForMob(mobId);
+                if (!soundGroups.isEmpty()) {
+                    mob.goalSelector.addGoal(3, new ReactToSoundGoal(mob, speed, range, soundGroups));
+                }
             }
         }
     }