+++ changelog.md
@@ -1,6 +1,7 @@
 ### Changelog
 
-#### Version 0.4.1
+#### Version 0.8.2
 
-- Organized the project structure to improve maintainability.
-- Added "Flee" mechanic, enhancing gameplay dynamics.
+
+- The configuration system has been **redesigned** for better clarity and organization. The configuration file paths have been changed to `config/ezvcsurvival/`, and the files are now named `voices.toml` and `sounds.toml`.
+- Guns of TAC with suppressors will no longer be heard by mobs.